requirements.txt
langchain
python-dotenv
langchain_groq
langchain_huggingface
hf_xet
langchain_community
streamlit
pypdf
faiss-cpu